fix(frontend): extract codemirror extension for ref update into hook
