+++ frontend/src/components/editor-page/editor-page-content.tsx
@@ -11,7 +11,7 @@ import { MotdModal } from '../common/motd-modal/motd-modal'
 import { CommunicatorImageLightbox } from '../markdown-renderer/extensions/image/communicator-image-lightbox'
 import { ExtensionEventEmitterProvider } from '../markdown-renderer/hooks/use-extension-event-emitter'
 import { AppBar, AppBarMode } from './app-bar/app-bar'
-import { ChangeEditorContentContextProvider } from './change-content-context/change-content-context'
+import { ChangeEditorContentContextProvider } from './change-content-context/codemirror-reference-context'
 import { EditorDocumentRenderer } from './editor-document-renderer/editor-document-renderer'
 import { EditorPane } from './editor-pane/editor-pane'
 import { useComponentsFromAppExtensions } from './editor-pane/hooks/use-components-from-app-extensions'
