+++ frontend/src/components/editor-page/editor-pane/hooks/use-handle-upload.tsx
@@ -7,7 +7,7 @@ import { uploadFile } from '../../../../api/media'
 import { getGlobalState } from '../../../../redux'
 import { supportedMimeTypes } from '../../../common/upload-image-mimetypes'
 import { useUiNotifications } from '../../../notifications/ui-notification-boundary'
-import type { ContentFormatter } from '../../change-content-context/change-content-context'
+import type { ContentFormatter } from '../../change-content-context/use-change-editor-content-callback'
 import { changeEditorContent } from '../../change-content-context/use-change-editor-content-callback'
 import { replaceInContent } from '../tool-bar/formatters/replace-in-content'
 import { replaceSelection } from '../tool-bar/formatters/replace-selection'
